+++ config/AppConfig/app.py
@@ -12,10 +12,10 @@ SHORT_NAME = "VP"
 DESCRIPTION = "VibraPilot browser automation desktop application by Vib Tools."
 TAGLINE = "Authorized browser automation desktop application."
 
-VERSION = "1.0.6.7"
+VERSION = "1.0.6.8"
 CREATED_DATE = "2026-08-07"
-RELEASE_DATE = "2026-08-08"
-UPDATED_DATE = "2026-08-08"
+RELEASE_DATE = "2026-08-09"
+UPDATED_DATE = "2026-08-09"
 STATUS = "development"
 RELEASE_CHANNEL = "production"